requirements.txt
absl-py
altair
annotated-types
arabic-reshaper
asgiref
asttokens
astunparse
attrs
backcall
backoff
bcrypt
beautifulsoup4
bleach
blinker
blis
cachetools
camel_tools
catalogue
certifi
chardet
charset-normalizer
click
cloudpathlib
colorama
comm
confection
contourpy
cycler
cymem
dataclasses
DateTime
debugpy
decorator
deep-translator
defusedxml
dill
Django
django-bootstrap-icons
django-bootstrap-v5
docopt
dotenv
dtaidistance
editdistance
emoji
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
env
et_xmlfile
executing
Faker
Farasa
fastjsonschema
filelock
flatbuffers
fonttools
frozendict
fsspec
future
gast
geocoder
gitdb
GitPython
google-pasta
grpcio
h5py
hijri-converter
html5lib
huggingface-hub
idna
ipykernel
ipynb
ipython
jedi
Jinja2
joblib
JPype1
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab_pygments
keras
kiwisolver
langcodes
language_data
libclang
loguru
lxml
lz4
marisa-trie
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
ml-dtypes
mplcursors
mplfinance
mpmath
muddler
multitasking
murmurhash
namex
narwhals
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nltk
numpy<2.0.0
openpyxl
opt_einsum
optree
packaging
pandas
pandas_ta
pandocfilters
parso
peewee
pickleshare
pillow
pip
pipreqs
platformdirs
plot-model
plotly
preshed
prompt_toolkit
protobuf
psutil
pure_eval
pyarrow
pyclustering
pydantic
pydantic_core
pydeck
pydot
Pygments
pyparsing
pyrsistent
python-bidi
python-crfsuite
python-dateutil
python-dotenv
pytz
PyYAML
pyzmq
ratelim
referencing
regex
reportlab
requests
rich
rpds-py
safetensors
scikit-learn
scipy
scrapfly-sdk
seaborn
setuptools
shellingham
six
sklearn-crfsuite
smart-open
smmap
soupsieve
spacy
spacy-legacy
spacy-loggers
sqlitecloud
sqlparse
srsly
stack-data
stopwords
streamlit
sympy
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow_intel
termcolor
thinc
threadpoolctl
tinycss2
tk
tokenizers
toml
torch
tornado
tqdm
traitlets
transformers
typer
typing_extensions
tzdata
urllib3
wasabi
watchdog
wcwidth
weasel
webencodings
Werkzeug
wheel
win32_setctime
wordcloud
wrapt
yarg
yfinance
zope.interface